Rename error serializer
