+++ src/ErrorResponse/Serializer/ComponentFactory/TypeFactory.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace App\ErrorResponse\ErrorSerializer\ComponentFactory;
+namespace App\ErrorResponse\Serializer\ComponentFactory;
 
 use App\ErrorResponse\ErrorInterface;
-use App\ErrorResponse\ErrorSerializer\Component;
-use App\ErrorResponse\ErrorSerializer\ComponentFactoryInterface;
+use App\ErrorResponse\Serializer\Component;
+use App\ErrorResponse\Serializer\ComponentFactoryInterface;
 
 class TypeFactory implements ComponentFactoryInterface
 {
